Types change: big pass on libstddjb and libunixonacid
 libdatastruct still missing, library still not functional

+++ src/include/skalibs/envalloc.h
@@ -3,13 +3,14 @@
 #ifndef ENVALLOC_H
 #define ENVALLOC_H
 
+#include <sys/types.h>
 #include <skalibs/genalloc.h>
 
 #define ENVALLOC_ZERO GENALLOC_ZERO
 
-extern int envalloc_make (genalloc *, unsigned int, char const *, unsigned int) ;
+extern int envalloc_make (genalloc *, size_t, char const *, size_t) ;
 extern int envalloc_uniq (genalloc *, char) ;
-extern int envalloc_merge (genalloc *, char const *const *, unsigned int, char const *, unsigned int) ;
+extern int envalloc_merge (genalloc *, char const *const *, size_t, char const *, size_t) ;
 extern int envalloc_0 (genalloc *) ;
 
 #endif
