+++ src/libstddjb/env_addmodif.c
@@ -1,11 +1,12 @@
 /* ISC license. */
 
+#include <sys/types.h>
 #include <skalibs/stralloc.h>
 #include <skalibs/env.h>
 
 int env_addmodif (stralloc *sa, char const *s, char const *t)
 {
-  unsigned int oldlen = sa->len ;
+  size_t oldlen = sa->len ;
   if (!s) return 1 ;
   if (!stralloc_cats(sa, s)) return 0 ;
   if ((t && (!stralloc_catb(sa, "=", 1) || !stralloc_cats(sa, t)))
