+++ src/libstddjb/timestamp_fmt.c
@@ -1,8 +1,9 @@
 /* ISC license. */
 
+#include <sys/types.h>
 #include <skalibs/tai.h>
 
-unsigned int timestamp_fmt (char *s, tain_t const *a)
+size_t timestamp_fmt (char *s, tain_t const *a)
 {
   *s = '@' ;
   return 1 + tain_fmt(s+1, a) ;
